Filter duplicated devices (#62)
* duplicated profile name fix

* rubocop fixes

* retry fix

* add fetch_device to device_client.rb

Fetch device list again after a new device registration on the Developer Portal

* Check the development and the ad-hoc profile's device list; if it's not up to date, then generate a new one

* add some comments

* pr updates

* rubocop fixes

* rubocop todo

* Check the profile device list only if the portal device list is not empty

* fix return value in the distribution_type_matches()

* fix: all_services_enabled() call argument fix: entitlement => profile

* pr updates

* clean

* ensure Bitrise test devices only in the Provisioning Profile

* PR: using return

* PR: using blank instead of nil check

* add activesupport to gemfile

* gemfile fix for rubocop

* add require 'active_support'

* remove the tvOS logic from test device check (Bitrise doesn't have tvOS device)

* activesupport gem removed; using `to_a.nil?` instead of the Rail's `blank?` solution

* rename filtered_test_device_udids -> test_device_udids in profile_client.rb

* add test device check for Xcode Managed Profiles

* fix the test_xcode_managed test workflow: GENERATE_PROFILES set to `no`

* check all the xcode managed profiles for the device liet

* changing the log from Print to debug in the device_list_up_to_date method

*  -

* temp: debug log for fetch profiles

* filter non Xcode Managed profiles if the generate set: no

* fix: reject Ad Hoc profiles if the distribution type is App Store

* feat: filter the duplicated test devices before registering them on the Developer Portal.

* replace the lambda with for each instead.

*  -

* PR: add some example Xcode Managed profile names in comment

* PR: check the Ad Hoc prefixes only for Xcode Managed profiles; add some profile name example in comment

* test: add test_xcode_managed_generate_enabled for testing Xcode Managed profiles

* typo fix

* TEMP LOG: inspect test device

* override == method for device

* Implement uniq method for test_device; It will compare two device by the uidid (without any special character)

* clean

* PR: uniq method renamed to filter_duplicated_devices

* RENAME: rename TestDevice to Device

* PR: clean filter_duplicated_devices method

* RENAME: rename test_device.rb to device.rb

* rename fix

* rename fix atempt 2
