+++ .rubocop_todo.yml
@@ -1,12 +1,12 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-02-11 16:03:44 +0100 using RuboCop version 0.58.2.
+# on 2019-02-19 13:51:31 +0100 using RuboCop version 0.58.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 31
+# Offense count: 33
 Metrics/AbcSize:
   Max: 65
 
@@ -21,7 +21,7 @@ Metrics/BlockLength:
 Metrics/ClassLength:
   Max: 331
 
-# Offense count: 15
+# Offense count: 16
 Metrics/CyclomaticComplexity:
   Max: 18
 
@@ -35,7 +35,7 @@ Metrics/MethodLength:
 Metrics/ParameterLists:
   Max: 8
 
-# Offense count: 12
+# Offense count: 13
 Metrics/PerceivedComplexity:
   Max: 22
 
@@ -44,6 +44,13 @@ Metrics/PerceivedComplexity:
 Style/BlockComments:
   Exclude:
     - 'spec/spec_helper.rb'
+
+# Offense count: 1
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: each, for
+Style/For:
+  Exclude:
+    - 'lib/autoprovision/device.rb'
 
 # Offense count: 1
 # Configuration parameters: MinBodyLength.
@@ -69,7 +76,7 @@ Style/RescueStandardError:
     - 'lib/autoprovision/profile_helper.rb'
     - 'step.rb'
 
-# Offense count: 227
+# Offense count: 229
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
